refactor: move bank to models submodule
