--- src/bank/bank_struct.rs
@@ -1,5 +1,0 @@
-// Interface for bank actions
-
-pub(crate) struct BankInterface {
-    // nothing
-}
