--- src/bank/calculate/mod.rs
@@ -1,3 +1,0 @@
-// Calculations around taxes.
-pub(crate) mod balance;
-pub(crate) mod fees;
