--- src/bank/deposit/mod.rs
@@ -1,2 +1,0 @@
-// Things that put money into the bank
-pub(crate) mod taxes;
