--- src/bank/mod.rs
@@ -1,7 +1,0 @@
-// Bank related things.
-pub(crate) mod bank_struct;
-pub(crate) mod calculate;
-pub(crate) mod deposit;
-pub(crate) mod modify;
-pub(crate) mod movement;
-pub(crate) mod withdrawal;
