--- src/bank/modify/mod.rs
@@ -1,2 +1,0 @@
-// Changes to the bank itself, like setting a tax rate.
-pub(crate) mod set_rate;
