--- src/bank/movement/mod.rs
@@ -1,2 +1,0 @@
-// Moving doints around should be done by the bank, since can check these methods more rigorously.
-pub(crate) mod move_doints;
