--- src/bank/withdrawal/mod.rs
@@ -1,2 +1,0 @@
-// Things that take money out of the bank
-pub(crate) mod ubi;
