+++ src/invocable/standard/casino/coin_flip.rs
@@ -4,11 +4,11 @@ use bigdecimal::{BigDecimal, FromPrimitive as _, Zero};
 use diesel::Connection;
 use log::{debug, warn};
 
-use crate::bank::bank_struct::BankInterface;
-use crate::bank::movement::move_doints::{DointTransfer, DointTransferParty, DointTransferReason};
 use crate::database::queries::user::get_doint_user;
 use crate::formatting::format_struct::FormattingHelper;
 use crate::guards;
+use crate::models::BankInterface;
+use crate::models::bank::transfer::{DointTransfer, DointTransferParty, DointTransferReason};
 use crate::types::serenity_types::{Context, Error};
 
 // a coin
