+++ src/lib.rs
@@ -1,7 +1,6 @@
 #![allow(clippy::result_large_err)]
 
 // Main can only see the discord side.
-mod bank;
 mod consent;
 mod database;
 pub mod discord;
@@ -12,6 +11,7 @@ mod invocable;
 mod jail;
 mod knob;
 mod tests;
+mod models;
 mod types;
 
 // Diesel related
