+++ src/models/bank/conversions.rs
@@ -1,0 +1,5 @@
+/// Convert a `tax_rate` into a percentage by dividing by 1000.
+#[inline]
+pub fn tax_rate_to_percentage(bank_tax_rate: i16) -> f64 {
+    f64::from(bank_tax_rate) / 1000.0
+}
