+++ src/models/bank/mod.rs
@@ -1,0 +1,6 @@
+pub mod transfer;
+pub mod data;
+pub mod conversions;
+pub mod taxes;
+pub mod fees;
+pub mod ubi;
