+++ src/models/bank/ubi.rs
@@ -2,7 +2,7 @@
 
 // Collect taxes from doint-holders.
 
-use crate::bank::bank_struct::BankInterface;
+use crate::models::BankInterface;
 use crate::database::tables::bank::BankInfo;
 use crate::schema::bank::dsl::bank;
 use crate::schema::users::dsl::users;
