+++ src/models/mod.rs
@@ -1,0 +1,3 @@
+pub mod bank;
+
+pub struct BankInterface {}
